refactor: move copy recursive to utils
