+++ vrc-get-vpm/src/add_package.rs
@@ -1,14 +1,13 @@
 use crate::structs::package::PackageJson;
-use crate::utils::{parse_hex_256, MapResultExt, PathBufExt};
+use crate::utils::{copy_recursive, parse_hex_256, MapResultExt, PathBufExt};
 use crate::{try_open_file, PackageInfo, PackageInfoInner};
 use futures::TryStreamExt;
 use indexmap::IndexMap;
 use reqwest::Client;
 use sha2::{Digest, Sha256};
-use std::collections::VecDeque;
 use std::io;
 use std::io::SeekFrom;
-use std::path::{Component, Path, PathBuf};
+use std::path::{Component, Path};
 use tokio::fs::{create_dir_all, remove_dir_all, File, OpenOptions};
 use tokio::io::{AsyncReadExt, AsyncSeekExt, AsyncWriteExt};
 use tokio_util::compat::{FuturesAsyncReadCompatExt, TokioAsyncReadCompatExt};
@@ -251,58 +250,4 @@ async fn add_local_package(
     let dest_folder = target_packages_folder.join(name);
     remove_dir_all(&dest_folder).await.ok();
     copy_recursive(package.to_owned(), dest_folder).await
-}
-
-async fn copy_recursive(src_dir: PathBuf, dst_dir: PathBuf) -> io::Result<()> {
-    // TODO: parallelize & speedup
-    let mut queue = VecDeque::new();
-    queue.push_front((src_dir, dst_dir));
-
-    while let Some((src_dir, dst_dir)) = queue.pop_back() {
-        let mut iter = tokio::fs::read_dir(src_dir).await?;
-        create_dir_all(&dst_dir).await?;
-        while let Some(entry) = iter.next_entry().await? {
-            let file_type = entry.file_type().await?;
-            let src = entry.path();
-            let dst = dst_dir.join(entry.file_name());
-
-            if file_type.is_symlink() {
-                // symlink: just copy
-                let symlink = tokio::fs::read_link(src).await?;
-                if symlink.is_absolute() {
-                    return Err(io::Error::new(
-                        io::ErrorKind::PermissionDenied,
-                        "absolute symlink detected",
-                    ));
-                }
-
-                #[cfg(unix)]
-                tokio::fs::symlink(dst, symlink).await?;
-                #[cfg(windows)]
-                {
-                    use std::os::windows::fs::FileTypeExt;
-                    if file_type.is_symlink_file() {
-                        tokio::fs::symlink_file(dst, symlink).await?;
-                    } else {
-                        assert!(file_type.is_symlink_dir(), "unknown symlink");
-                        tokio::fs::symlink_dir(dst, symlink).await?;
-                    }
-                }
-                #[cfg(not(any(unix, windows)))]
-                return Err(io::Error::new(
-                    io::ErrorKind::Unsupported,
-                    "platform without symlink detected",
-                ));
-            } else if file_type.is_file() {
-                tokio::fs::copy(src, dst).await?;
-            } else if file_type.is_dir() {
-                //copy_recursive(&src, &dst).await?;
-                queue.push_front((src, dst));
-            } else {
-                panic!("unknown file type: none of file, dir, symlink")
-            }
-        }
-    }
-
-    Ok(())
 }
